Use vanilla CSS to automatically place sponsor logos
This is using css grids rather than tailwind's `flex-row` property.

The `grid-tempate-colums` style on line 84 ensures that the smallest
width of an image can be 250px. if there is extra space, `auto-fit` will
ensure the images take as much space as possible on the row. otherwise,
the element will overflow to a new row.

for additional info, see [this
page](https://css-tricks.com/auto-sizing-columns-css-grid-auto-fill-vs-auto-fit/)
(or message me)